requirements.txt
PyQt6>=6.6.1
google-genai>=0.2.0
Pillow>=10.2.0
cryptography>=42.0.0
markdown>=3.5.0
